% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addendum_data.R
\docType{data}
\name{data_example_long_1d}
\alias{data_example_long_1d}
\title{A simulated dataset with univariate data}
\format{
An object of class \code{list} of length 9.
}
\usage{
data(data_example_long_1d)
}
\value{
A list of the following components
\item{$data_matrix_IC}{The data matrix for IC data.}
\item{$time_matrix_IC}{The time matrix for IC data.}
\item{$nobs_IC}{Number of observations for each IC process.}
\item{$data_matrix_OC}{The data matrix for OC data.}
\item{$time_matrix_OC}{The time matrix for OC data.}
\item{$nobs_OC}{Number of observations for each OC process.}
\item{$design_interval}{The design interval.}
\item{$n_time_units}{Number of time units in the design interval.}
\item{$time_unit}{The time unit.}
}
\description{
A simulated univariate longitudinal dataset for demonstration.
}
\details{
Data Example: Univariate Longitudinal Data
}
\examples{
data(data_example_long_1d)
}
\keyword{datasets}
