% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addendum_data.R
\docType{data}
\name{data_example_long_surv}
\alias{data_example_long_surv}
\title{A simulated dataset with longitudinal and survival data}
\format{
An object of class \code{list} of length 15.
}
\usage{
data(data_example_long_surv)
}
\value{
A list of the following components
\item{$data_array_IC}{The data array for IC data.}
\item{$time_matrix_IC}{The time matrix for IC data.}
\item{$nobs_IC}{Number of observations for each IC process.}
\item{$starttime_IC}{Start time of monitoring for IC processes.}
\item{$survtime_IC}{End time of monitoring for IC processes.}
\item{$survevent_IC}{Survival events of IC processes.}
\item{$data_array_OC}{The data array for OC data.}
\item{$time_matrix_OC}{The time matrix for OC data.}
\item{$nobs_OC}{Number of observations for each OC process.}
\item{$starttime_OC}{Start time of monitoring for OC processes.}
\item{$survtime_OC}{End time of monitoring for OC processes.}
\item{$survevent_OC}{Survival events of OC processes.}
\item{$design_interval}{The design interval.}
\item{$n_time_units}{Number of time units in the design interval.}
\item{$time_unit}{The time unit.}
}
\description{
A simulated univariate longitudinal dataset for demonstration.
}
\details{
Data Example: Longitudinal and Survival Data
}
\examples{
data(data_example_long_surv)
}
\keyword{datasets}
