% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addendum_data.R
\docType{data}
\name{data_stroke}
\alias{data_stroke}
\title{A real data example on stroke}
\format{
An object of class \code{list} of length 8.
}
\usage{
data(data_stroke)
}
\value{
A list of the following components
\item{$systolic_ctrl}{A matrix of systolic blood pressures for controls. The [i,j] element is the jth observation of the ith control.}
\item{$diastolic_ctrl}{A matrix of diastolic blood pressures for controls. The [i,j] element is the jth observation of the ith control.}
\item{$cholesterol_ctrl}{A matrix of cholesterol levels for controls. The [i,j] element is the jth observation of the ith control.}
\item{$age_ctrl}{A matrix of the age of observations for controls. The [i,j] element is the age of jth observation for the ith control.}
\item{$systolic_case}{A matrix of systolic blood pressures for cases. The [i,j] element is the jth observation of the ith case.}
\item{$diastolic_case}{A matrix of diastolic blood pressures for cases. The [i,j] element is the jth observation of the ith case.}
\item{$cholesterol_case}{A matrix of cholesterol levels for cases. The [i,j] element is the jth observation of the ith case.}
\item{$age_case}{A matrix of the age of observations for cases. The [i,j] element is the age of jth observation for the ith case.}
}
\description{
In this dataset, there are 27 subjects with stroke and 1028 subjects without stroke. 
Three risk factors, systolic blood pressures, diastolic blood pressures, cholesterol levels, 
are collected over time at different ages.
}
\details{
Real Data Example: Stroke Data
}
\examples{
data(data_stroke)
}
\keyword{datasets}
